Rename integration namespace
Make it clearer that these tests test some unit + real database
