+++ tests/ManageCourses.Tests/DbIntegration/UserServiceTests.cs
@@ -8,11 +8,11 @@ using GovUk.Education.ManageCourses.Api.Services;
 using GovUk.Education.ManageCourses.Api.Services.Email;
 using GovUk.Education.ManageCourses.Api.Services.Email.Model;
 using GovUk.Education.ManageCourses.Domain.Models;
-using GovUk.Education.ManageCourses.Tests.Integration.DatabaseAccess;
+using GovUk.Education.ManageCourses.Tests.DbIntegration.DatabaseAccess;
 using Moq;
 using NUnit.Framework;
 
-namespace GovUk.Education.ManageCourses.Tests.Integration
+namespace GovUk.Education.ManageCourses.Tests.DbIntegration
 {
     [TestFixture]
     [Category("Integration")]
